Level so Far
It adds in ProGrid as it helps in positioning things within the level and making it so things don't clash in UV's.

It also changed something with ShaderOutlineBlue. Not too sure what it was but it just added itself onto the list of changes.

As well as the first sections of the PoC level. More to be added later but for now it is barebones.