PoC Work - AC
The scene is done for now. Lighting needs to be re-baked so that the glitches stop appearing, but a more powerful processor is needed to do so.

There are objects that need making for the building to make it seem less empty. If there are any issues with the lighting or processing speed when changes are made, contact me and I will take you through the steps to rectify it.

De-activate the probuilder menu when you open the scene to ensure that you don't accidenly modify anything.