Finalised basement positions
Only walls done so far, the vents will be done next with the placeholder guards being added in with security cameras and ID doors being moved last.

Work will also be done for the sequenced button placements and a start for designing the waypoint needed for the player to properly navigate the level without any external help.