Updated Build *Read Description
The roof has been finished in the main stage. Took a very long time so if there is an issue use probuilder to change it.

There has been an issue with the spawning in of the code hints for the keypad. Not sure what it is but the spawns don't quite work.

There is a missing link for the key pickup on the door, not sure what its supposed to be.

Key placement is finalised, hassan's prison code in the GM_GameManager needs to be filled in and guard positions need to be added into the stage.

We need a teleport script, something to allow the player to pickup the diamond when they have talked to toni, the end zones need to be decided as well as a few more car models being needed to fill out a car park.

Also, Laundry shoot doesn't actually exist and now that the basement is gone, its no longer needed.